Rename signaling to back-end & update static serving to support SPA
