+++ package.json
@@ -2,12 +2,12 @@
   "name": "watchme",
   "version": "1.0.0",
   "description": "share your screen without headaches",
-  "main": "signaling/app.js",
+  "main": "back-end/app.js",
   "scripts": {
-    "install": "cd signaling; npm install",
-    "start": "cd signaling; npm start",
-    "dev": "cd signaling; npm run dev",
-    "test": "cd signaling; npm test"
+    "install": "cd back-end; npm install; cd ../front-end; npm install",
+    "start": "cd front-end; npm run build; cd ../back-end; npm start",
+    "dev": "cd front-end; npm run build; cd ../back-end; npm run dev",
+    "test": "cd back-end; npm test"
   },
   "repository": {
     "type": "git",
